remove quantum copilot
