--- vscode/src/copilot/webview/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "CommonJS",
-    "target": "ES2022",
-    "noEmit": true,
-    "lib": ["DOM", "ES2022"],
-    "rootDir": ".",
-    "strict": true /* enable all strict type-checking options */,
-    "jsx": "react-jsx",
-    "jsxImportSource": "preact",
-    "isolatedModules": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "types": []
-  }
-}
