--- vscode/src/copilot/webview/types.d.ts
@@ -1,5 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-// This module has no types available, so we declare it as any.
-declare module "highlightjs-copy";
